Minor Fixes

+++ stdplugins/tagall.py
@@ -1,20 +1,23 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+"""Command: .tagall optional_text_to_tag_in
+You can use it in reply to a message or directly in a new message.
+  © [cHAuHaN](http://t.me/amnd33p)"""
 from telethon import events
 from uniborg.util import admin_cmd
 
-
-@borg.on(admin_cmd("tagall"))
+@borg.on(admin_cmd("tagall($| .*)"))
 async def _(event):
     if event.fwd_from:
         return
     try:
     	mentions = event.pattern_match.group(1)
     except Exception as e:
-    	mentions = "You"
+    	mentions = "..."
     chat = await event.get_input_chat()
     async for x in borg.iter_participants(chat, 100):
         mentions += f"[\u2063](tg://user?id={x.id})"
-    await event.reply(mentions)
-    await event.delete()
+    if event.reply_to_msg_id:
+        previous_message = await event.get_reply_message()
+        await previous_message.reply(mentions)
+    else:
+        await event.reply(mentions)
+        await event.delete()
